Add attempt at single-PRNG SIMD vars
The theory is that there are less sequential instructions necessary to
compute a single COMBIT PRNG instance by packing the gears in a single
SIMD variable.

In practice, we cannot do this at 64-bit gears, because AVX2 only
supports up to 256 bits, which only allow 4 gears.
There are no current COMBIT variant that beats 32 TiB of PractRand with
only 4 gears. The best we have is 5. So close!

So this means halving the output to a 32-bit output, which has inferior
performance to the computing-multiple-PRNGs-in-parallel-in-SIMD
approach.

I keep it around for two reasons:

1. Maybe further study will yield a 4-gear COMBIT variant that beats
   PractRand.
2. AVX-512 supports up to 8 64-bit gears.